Reminder: edit all weapons' Traits stat
This update removed an empty rule from Models, and changed the name of a rule in Model/Weapon Properties (for clarity).

That last rule was changed to add the assets' designations to any parent, not just Models; that means that all weapon assets will need to be edited to remove existing strings in the Traits stat.